pi/requirements.txt
ads1015==0.0.5
boto3==1.10.17
botocore==1.13.17
certifi==2019.9.11
chardet==3.0.4
docutils==0.15.2
enviroplus==0.0.1
i2cdevice==0.0.6
idna==2.8
jmespath==0.9.4
ltr559==0.1.0
numpy==1.17.4
Pillow==6.2.1
pimoroni-bme280==0.0.2
pkg-resources==0.0.0
pms5003==0.0.4
pyserial==3.4
python-dateutil==2.8.0
requests==2.22.0
RPi.GPIO==0.7.0
s3transfer==0.2.1
six==1.13.0
smbus==1.1.post2
spidev==3.4
ST7735==0.0.3
urllib3==1.25.7
